Use separate class for Schema
